+++ src/main/java/bridge/view/OutputView.java
@@ -16,14 +16,6 @@ public class OutputView {
         System.out.println("다리 건너기 게임을 시작합니다.");
     }
 
-    public void printInputBridgeLengthMessage() {
-        System.out.println("다리의 길이를 입력해주세요.");
-    }
-
-    public void printMovingMessage() {
-        System.out.println("이동할 칸을 선택해주세요. (위: U, 아래: D)");
-    }
-
     /**
      * 현재까지 이동한 다리의 상태를 정해진 형식에 맞춰 출력한다.
      * <p>
@@ -51,10 +43,6 @@ public class OutputView {
         System.out.println("[ " + upString + " ]");
         System.out.println("[ " + downString + " ]");
         // O | O | X
-    }
-
-    public void printGameCommandMessage() {
-        System.out.println("게임을 다시 시도할지 여부를 입력해주세요. (재시도: R, 종료: Q)");
     }
 
     /**
